Split file into multiple files
Based on feedback from @Sneh1999 and @zacshowa:
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279797524
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#pullrequestreview-3084968050

With the added functionality and types introduced for multi worker queue processing
the solution has become somewhat large and can make it more difficult to navigate
the implementation.  The suggestion has been put forth to separate the large file
into separate files.

This change separates each of the primary types into their own files for isolated
references where possible.  The errors have been kept next to their place of utilization
for faster lookup.
